UDPクラスの更新 ＊更新内容＊ charからint16_tに変更することで通信量を50%削減 各ノード内での配列dataの宣言型を変更
