+++ f7_udp_cpp/src/NR25_DR.cpp
@@ -20,7 +20,7 @@ RRST NHK2025
 std::string udp_ip = "192.168.0.218"; // 送信先IPアドレス、宛先マイコンで設定したIPv4アドレスを指定
 int udp_port = 5000;                  // 送信元ポート番号、宛先マイコンで設定したポート番号を指定
 
-std::vector<int> data = {0, -1, -1, -1, -1, -1, 0, 0, 0}; // 1~5番を電磁弁制御に転用中（-1 or 1）
+std::vector<int16_t> data = {0, -1, -1, -1, -1, -1, 0, 0, 0}; // 1~5番を電磁弁制御に転用中（-1 or 1）
 
 // 各機構のシーケンスを格納するクラス
 class Action {
