+++ f7_udp_cpp/src/NR25_MR_SwerveDrive.cpp
@@ -71,7 +71,7 @@ int SERVO4_CAL = 0;
 std::string udp_ip = "192.168.8.215"; // 送信先IPアドレス、宛先マイコンで設定したIPv4アドレスを指定
 int udp_port = 5000;                  // 送信元ポート番号、宛先マイコンで設定したポート番号を指定
 
-std::vector<int> data = {0, 0, 0, 0, 0, 0, 0, 0, 0};
+std::vector<int16_t> data = {0, 0, 0, 0, 0, 0, 0, 0, 0};
 
 class PS4_Listener : public rclcpp::Node {
 public:
