+++ f7_udp_cpp/src/NR25_PID.cpp
@@ -26,7 +26,7 @@ PID on ROS
 std::string udp_ip = "192.168.8.217"; // 送信先IPアドレス、宛先マイコンで設定したIPv4アドレスを指定
 int udp_port = 5000;                  // 送信元ポート番号、宛先マイコンで設定したポート番号を指定
 
-std::vector<int> data = {0, 0, 0, 0, 0, 0, -1, -1, -1}; // 7~9番を電磁弁制御に転用中（-1 or 1）
+std::vector<int16_t> data = {0, 0, 0, 0, 0, 0, -1, -1, -1}; // 7~9番を電磁弁制御に転用中（-1 or 1）
 
 float v[5] = {0.0, 0.0, 0.0, 0.0, 0.0};
 // float d[5] = {0.0, 0.0, 0.0, 0.0, 0.0};
